+++ src/screens/famCup/createScore.tsx
@@ -9,9 +9,9 @@ import { colors } from "@theme";
 import { officeStyles } from "@styles";
 import { FloatingValidateBtn } from "components/validateButton";
 import { PointInputController } from "components/pointInput";
-import { PointsFieldNames } from "@types";
 import { usePoint } from "firebase/api";
 import { DateComponent } from "components/dateScoreInput";
+import { PointsFieldNames } from "types/point.type";
 
 export default function CreateScoreScreen({ navigation }: CreateScoreProps) {
   const { createPoint } = usePoint();
