+++ src/types/date.type.ts
@@ -1,0 +1,10 @@
+export type DateType = {
+  start: string;
+  end: string;
+};
+
+export type DatePickerValues = {
+  showStart?: boolean;
+  showEnd?: boolean;
+  mode: "date" | "time";
+};
