+++ src/types/student.type.ts
@@ -4,5 +4,17 @@ export type Student = {
   lastName: string;
   firstName: string;
   adhesion?: string[];
-  memberOf?: string[];
+};
+
+export type FirestoreStudent = {
+  mail: string;
+  lastName: string;
+  firstName: string;
+  adhesion?: string[];
+};
+
+export type CreateStudentArgs = {
+  firstName: string;
+  lastName: string;
+  mail: string;
 };
